code_reviewer/requirements.txt
gitpython
openai
pyright
